Create campaigner and member facing webpack bundles
