+++ app/views/layouts/application.html.slim
@@ -5,6 +5,7 @@ html
     meta name="viewport" content="width=device-width, initial-scale=1.0"
 
     = stylesheet_link_tag    "application", media: 'all'
+    = stylesheet_link_tag *webpack_asset_paths('campaigners', extension: 'css') unless Rails.env.test?
     = javascript_include_tag "application"
 
     = csrf_meta_tags
@@ -26,4 +27,4 @@ html
     = render 'shared/flash'
 
     = yield
-    = javascript_include_tag *webpack_asset_paths('components', extension: 'js') unless Rails.env.test?
+    = javascript_include_tag *webpack_asset_paths('campaigners', extension: 'js') unless Rails.env.test?
