+++ app/views/layouts/member_facing.slim
@@ -7,7 +7,7 @@ html lang="#{@page.language.try(:code).try(:downcase)}"
     - if Settings.facebook_app_id.present?
       meta property='fb:app_id' content="#{Settings.facebook_app_id}"
     = stylesheet_link_tag 'member-facing'
-    = stylesheet_link_tag *webpack_asset_paths('components', extension: 'css') unless Rails.env.test?
+    = stylesheet_link_tag *webpack_asset_paths('members', extension: 'css') unless Rails.env.test?
     = csrf_meta_tags
     = favicon_link_tag 'images/favicon.ico'
     = metamagic
@@ -27,7 +27,7 @@ html lang="#{@page.language.try(:code).try(:downcase)}"
     = render partial: 'shared/page_object'
     = render partial: 'shared/facebook_pixel' unless Rails.env.test?
     = javascript_include_tag "https://cdn.polyfill.io/v2/polyfill.min.js?features=Intl.~locale.en,Intl.~locale.de,Intl.~locale.es,Intl.~locale.fr"
-    = javascript_include_tag *webpack_asset_paths('components', extension: 'js') unless Rails.env.test?
+    = javascript_include_tag *webpack_asset_paths('members', extension: 'js') unless Rails.env.test?
     = render partial: 'shared/google_analytics_snippet'
 
     script type="text/javascript" charset="utf-8" async=true src="//c.shpg.org/99/sp.js"
