+++ package.json
@@ -27,6 +27,8 @@
     "jquery": "^3.1.1",
     "js-cookie": "^2.1.3",
     "lodash": "^4.16.6",
+    "preact": "^7.2.1",
+    "preact-compat": "^3.14.3",
     "react": "^15.4.2",
     "react-dom": "^15.3.2",
     "react-intl": "^2.2.3",
@@ -49,6 +51,7 @@
     "browserify-incremental": "^3.0.1",
     "case-sensitive-paths-webpack-plugin": "1.1.4",
     "chalk": "1.1.3",
+    "compression-webpack-plugin": "^0.3.2",
     "connect-history-api-fallback": "1.3.0",
     "cross-spawn": "4.0.2",
     "css-loader": "^0.26.1",
